% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRPClustering.R
\name{crp_graph_2d}
\alias{crp_graph_2d}
\title{CRP clustering visualization}
\usage{
crp_graph_2d(data = data, z_result = z_result)
}
\arguments{
\item{data}{: a matrix of data for clustering. Row is each data_i and column is dimensions of each data_i.}

\item{z_result}{: an array denotes the number of a cluster for each data_i and it is the output of the method "crp_gibbs".}
}
\description{
CRP clustering visualization
}
