% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{crp_plot}
\alias{crp_plot}
\title{CRP Plot Matrix Visualization For Predicting}
\usage{
crp_plot(data = data, predict = predict)
}
\arguments{
\item{data}{: a data.frame of data for clustering. Row is each data_i and column is dimensions of each data_i.}

\item{predict}{: an array returned from crp_predict method.}
}
\description{
CRP Plot Matrix Visualization For Predicting
}
