% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{crp_plot_z}
\alias{crp_plot_z}
\title{Visualization Probability of the Clusters that Data i Join in.}
\usage{
crp_plot_z(i, result = result)
}
\arguments{
\item{i}{: a number of each data i.}

\item{result}{: a list returned from crp_train method. }
}
\description{
Visualization Probability of the Clusters that Data i Join in.
}
