% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{crp_train}
\alias{crp_train}
\title{Markov chain Monte Carlo methods for CRP Clustering}
\usage{
crp_train(data, mu = c(0, 0), sigma_table = 1, alpha = 1, ro_0 = 1,
  burn_in = 100, iteration = 1000)
}
\arguments{
\item{data}{: a matrix of data for clustering. Row is each data_i and column is dimensions of each data_i.}

\item{mu}{: a vector of center points of data. If data is 3 dimensions, a vector of 3 elements like c(2,4,2).}

\item{sigma_table}{: a numeric of CRP variance.}

\item{alpha}{: a numeric of a CRP concentrate rate.}

\item{ro_0}{: a numeric of a CRP mu change rate.}

\item{burn_in}{:  an iteration integer of burn in.}

\item{iteration}{: an iteration integer.}
}
\value{
result : an array expresses cluster number and joined data number and cluster's mean and variance matrix.
}
\description{
Markov chain Monte Carlo methods for CRP Clustering
}
\examples{
data <- matrix(c(1.8,1.9,2.1,2.5,5.6,5.2,6,6.1), 4, 2)
result <- crp_train(
                      data,
                      mu=c(0,0),
                      sigma_table=1,
                      alpha=1,
                      ro_0=1,
                      burn_in=100,
                      iteration=1000
                     )
}
