% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocument.R
\docType{data}
\name{data}
\alias{data}
\title{Generated from three normal distributions.}
\format{A data frame with 1000 rows and 2 dimensions}
\usage{
data
}
\description{
Generated from three normal distributions.
}
\keyword{datasets}
