% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{gibbs}
\alias{gibbs}
\title{Markov chain Monte Carlo methods. This method is an inner method.}
\usage{
gibbs(data_length, dim, data_raw,data_raw_result, alpha, burn_in, iteration)
}
\arguments{
\item{data_length}{:data size.}

\item{dim}{:data dimention.}

\item{data_raw}{:scaled data.}

\item{data_raw_result}{:raw data.}

\item{alpha}{: a numeric of a CRP concentrate rate.}

\item{burn_in}{: an iteration integer of burn in.}

\item{iteration}{: an iteration integer.}

}

\value{
result : an list has three elements. The "clusters" is cluster number and joined data number and cluster's mean and variance matrix. The "max" is the cluster number for data i join in. The "z" is the iteration history for an each data i join in clusters.
}
\description{
Inner method call from crp_train method.
}
