% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRPClustering.R
\name{crp_gibbs}
\alias{crp_gibbs}
\title{Markov chain Monte Carlo methods for CRP Clustering}
\usage{
crp_gibbs(data, mu = c(0, 0), sigma = 0.5, sigma_table = 12, alpha = 1,
  ro_0 = 0.1, burn_in = 10, iteration = 100)
}
\arguments{
\item{data}{: a matrix of data for clustering. Rows is data_i and cols is dimensions of data.}

\item{mu}{: numbers of center points of data. if data is 3 dimensions, a vector of 3 elements like =c(2,4,2).}

\item{sigma}{: a number of data variance.}

\item{sigma_table}{: a number of CRP variance.}

\item{alpha}{: a number of CRP concentrate rate.}

\item{ro_0}{: a number of CRP mu change rate.}

\item{burn_in}{:  iteration number of burn in.}

\item{iteration}{: iteration number.}
}
\value{
z_result : a vector express cluster a number for each data i.
}
\description{
Markov chain Monte Carlo methods for CRP Clustering
}
\examples{
z_result <- crp_gibbs(matrix(c(0.1,0.1,0.2,0.2,0.3,0.3,1.4,1.4,1.5,1.5),2,2),
 mu=c(0,0), sigma=0.5, sigma_table=12, alpha=1, ro_0=0.1, burn_in=10, iteration=100)
}
