% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRPClustering.R
\name{crp_gibbs}
\alias{crp_gibbs}
\title{Markov chain Monte Carlo methods for CRP Clustering}
\usage{
crp_gibbs(path = "./data/data.csv", mu = c(0, 0), sigma = 0.5,
  sigma_table = 12, burn_in = 10, iteration = 100)
}
\arguments{
\item{path}{: a data path and data. data.csv is a comma separate file. And cols is a dimension of data.}

\item{mu}{: a floating-point type of center points of data. if data is 3 dimensions, a vector of 3 elements like =c(2,4,2)}

\item{sigma}{: a floating-point type of data variance.}

\item{sigma_table}{: a floating-point type of table position variance.}

\item{burn_in}{:  iteration numbers of burn in.}

\item{iteration}{: iteration numbers.}
}
\value{
z_result : a vector and express cluster numbers for each data i
}
\description{
Markov chain Monte Carlo methods for CRP Clustering
}
