% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRPClustering.R
\name{crp_graph_2d}
\alias{crp_graph_2d}
\title{CRP Clustering Visualization}
\usage{
crp_graph_2d(path = "./data/data.csv", z_result)
}
\arguments{
\item{path}{: a data path and data. data.csv is a comma separate file. And cols is a dimension of data.}

\item{z_result}{: the output of the method "crp_gibbs".}
}
\description{
CRP Clustering Visualization
}
