% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRPClustering.R
\name{crp_graph_2d}
\alias{crp_graph_2d}
\title{CRP Clustering Visualization}
\usage{
crp_graph_2d(data, z_result)
}
\arguments{
\item{data}{: a matrix of data for Clustering. Rows are data_i and cols are dimensions of data.}

\item{z_result}{: a vector expresses cluster numbers for each data i. And the output of the method "crp_gibbs".}
}
\description{
CRP Clustering Visualization
}
\examples{
crp_graph_2d(matrix(c(0.1,0.1,0.2,0.2,0.3,0.3,1.4,1.4,1.5,1.5),2,2), c(1,1,2,2,2))
}
