% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{crp_predict}
\alias{crp_predict}
\title{Predict The Clusters' Probability that Data Joins.}
\usage{
crp_predict(data = data, result = result)
}
\arguments{
\item{data:}{a matrix of data for clustering. Row is each data_i and column is dimensions of each data_i.}

\item{result:}{an array expresses the return of crp_train method.}
}
\value{
predict:an array expresses a joined cluster number and joined probability for each cluster.
}
\description{
Predict The Clusters' Probability that Data Joins.
}
